tests/wpt/web-platform-tests/navigation-api/navigation-methods/navigate-replace-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation (which would defeat the point of the test).
  window.onload = () => t.step_timeout(() => {
    let start_history_length = history.length;
    let start_entry_top = navigation.currentEntry;
    let start_entry_iframe_id = i.contentWindow.navigation.currentEntry.id;
    let start_entry_iframe_key = i.contentWindow.navigation.currentEntry.key;

    i.contentWindow.navigation.navigate("?1", { history: "replace" });
    i.onload = t.step_func_done(() => {
      assert_equals(history.length, start_history_length);

      assert_equals(navigation.entries().length, 1);
      assert_equals(navigation.currentEntry, start_entry_top);

      assert_equals(i.contentWindow.navigation.entries().length, 1);
      let iframe_entry = i.contentWindow.navigation.currentEntry;
      assert_not_equals(start_entry_iframe_id, iframe_entry.id);
      assert_equals(start_entry_iframe_key, iframe_entry.key);
    });
  }, 0);
}, "navigate() with history: 'replace' option");
</script>
