LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/navigate-replace-cross-document-expected.txt



FAIL navigate() with history: 'replace' option assert_equals: expected 2 but got 3
